Create BaseSetOptions that SortedSet and Set can extend
