+++ geode-apis-compatible-with-redis/src/integrationTest/resources/org/apache/geode/codeAnalysis/sanctionedDataSerializables.txt
@@ -34,11 +34,11 @@ org/apache/geode/redis/internal/data/RedisString,2
 toData,23
 fromData,23
 
-org/apache/geode/redis/internal/executor/sortedset/SortedSetOptions,2
+org/apache/geode/redis/internal/executor/BaseSetOptions,2
 fromData,14
 toData,9
 
 org/apache/geode/redis/internal/executor/string/SetOptions,2
-fromData,34
-toData,29
+fromData,27
+toData,27
