+++ geode-apis-compatible-with-redis/src/main/resources/org/apache/geode/redis/internal/sanctioned-geode-apis-compatible-with-redis-serializables.txt
@@ -9,8 +9,6 @@ org/apache/geode/redis/internal/delta/DeltaType,false
 org/apache/geode/redis/internal/executor/CommandFunction,true,-1302506316316454732
 org/apache/geode/redis/internal/executor/SingleResultRedisFunction,true,3239452234149879302
 org/apache/geode/redis/internal/executor/key/RenameFunction,true,7047473969356686453
-org/apache/geode/redis/internal/executor/sortedset/SortedSetOptions$Exists,false
-org/apache/geode/redis/internal/executor/string/SetOptions$Exists,false
 org/apache/geode/redis/internal/netty/CoderException,true,4707944288714910949
 org/apache/geode/redis/internal/netty/RedisCommandParserException,true,4707944288714910949
 org/apache/geode/redis/internal/pubsub/PubSubImpl$1,false,this$0:org/apache/geode/redis/internal/pubsub/PubSubImpl
